# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "LINUX DO PAY",
        "contact": {},
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/user/pay-key": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdatePayKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseAny"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "user.UpdatePayKeyRequest": {
            "type": "object",
            "required": [
                "pay_key"
            ],
            "properties": {
                "pay_key": {
                    "type": "string",
                    "maxLength": 6
                }
            }
        },
...
        "util.ResponseAny": {
            "type": "object",
            "properties": {
                "data": {},
                "error_msg": {
                    "type": "string",
                    "example": ""
                }
            }
        }
    }
}